requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
ollama>=0.1.6
groq>=0.5.0
huggingface_hub>=0.20.0
pypdf>=4.0.0
pickle5; python_version < '3.8'
